Replace `getResolutionRules()` with `setResolveMode(ResolveMode)`
