+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/ivyresolve/StartParameterResolutionOverride.java
@@ -16,12 +16,8 @@
 package org.gradle.api.internal.artifacts.ivyservice.ivyresolve;
 
 import org.gradle.StartParameter;
-import org.gradle.api.Action;
-import org.gradle.api.internal.artifacts.cache.ArtifactResolutionControl;
-import org.gradle.api.internal.artifacts.cache.DependencyResolutionControl;
-import org.gradle.api.internal.artifacts.cache.ModuleResolutionControl;
-import org.gradle.api.internal.artifacts.cache.ResolutionRules;
 import org.gradle.api.artifacts.component.ModuleComponentIdentifier;
+import org.gradle.api.internal.artifacts.configurations.ResolutionStrategyInternal.ResolveMode;
 import org.gradle.api.internal.artifacts.ivyservice.resolutionstrategy.ExternalResourceCachePolicy;
 import org.gradle.api.internal.artifacts.repositories.resolver.MetadataFetchingCost;
 import org.gradle.api.internal.component.ArtifactType;
@@ -47,7 +43,10 @@ import javax.annotation.Nullable;
 import java.io.IOException;
 import java.net.URI;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
+
+import static org.gradle.api.internal.artifacts.configurations.ResolutionStrategyInternal.ResolveMode.DEFAULT;
+import static org.gradle.api.internal.artifacts.configurations.ResolutionStrategyInternal.ResolveMode.OFFLINE;
+import static org.gradle.api.internal.artifacts.configurations.ResolutionStrategyInternal.ResolveMode.REFRESH_DEPENDENCIES;
 
 public class StartParameterResolutionOverride {
     private final StartParameter startParameter;
@@ -56,40 +55,8 @@ public class StartParameterResolutionOverride {
         this.startParameter = startParameter;
     }
 
-    public void addResolutionRules(ResolutionRules resolutionRules) {
-        if (startParameter.isOffline()) {
-            resolutionRules.eachDependency(new Action<DependencyResolutionControl>() {
-                public void execute(DependencyResolutionControl dependencyResolutionControl) {
-                    dependencyResolutionControl.useCachedResult();
-                }
-            });
-            resolutionRules.eachModule(new Action<ModuleResolutionControl>() {
-                public void execute(ModuleResolutionControl moduleResolutionControl) {
-                    moduleResolutionControl.useCachedResult();
-                }
-            });
-            resolutionRules.eachArtifact(new Action<ArtifactResolutionControl>() {
-                public void execute(ArtifactResolutionControl artifactResolutionControl) {
-                    artifactResolutionControl.useCachedResult();
-                }
-            });
-        } else if (startParameter.isRefreshDependencies()) {
-            resolutionRules.eachDependency(new Action<DependencyResolutionControl>() {
-                public void execute(DependencyResolutionControl dependencyResolutionControl) {
-                    dependencyResolutionControl.cacheFor(0, TimeUnit.SECONDS);
-                }
-            });
-            resolutionRules.eachModule(new Action<ModuleResolutionControl>() {
-                public void execute(ModuleResolutionControl moduleResolutionControl) {
-                    moduleResolutionControl.cacheFor(0, TimeUnit.SECONDS);
-                }
-            });
-            resolutionRules.eachArtifact(new Action<ArtifactResolutionControl>() {
-                public void execute(ArtifactResolutionControl artifactResolutionControl) {
-                    artifactResolutionControl.cacheFor(0, TimeUnit.SECONDS);
-                }
-            });
-        }
+    public ResolveMode getResolveMode() {
+        return startParameter.isOffline()? OFFLINE : startParameter.isRefreshDependencies()? REFRESH_DEPENDENCIES : DEFAULT;
     }
 
     public ModuleComponentRepository overrideModuleVersionRepository(ModuleComponentRepository original) {
